Move edge probing tasks to a separate .py file
