+++ src/tasks/__init__.py
@@ -1,5 +1,6 @@
 # Import task definitions to register their tasks.
 from . import tasks
+from . import edge_probing
 
 # REGISTRY needs to be available to modules within this package,
 # but we also import it here to make it available at the package level.
