+++ src/tasks/tasks.py
@@ -30,9 +30,7 @@ from allennlp.data import Instance, Token
 from allennlp.data.fields import TextField, LabelField, \
     SpanField, ListField, MetadataField
 from ..allennlp_mods.numeric_field import NumericField
-from ..allennlp_mods.multilabel_field import MultiLabelField
 
-from ..utils import serialize
 from ..utils import utils
 from ..utils.utils import load_tsv, process_sentence, truncate, load_diagnostic_tsv
 
@@ -43,7 +41,7 @@ UNK_TOK_ATOMIC = "UNKNOWN"  # an unk token that won't get split by tokenizers
 
 from .registry import register_task, REGISTRY  # global task registry
 
-def _sentence_to_text_field(sent: Sequence[str], indexers: Any):
+def sentence_to_text_field(sent: Sequence[str], indexers: Any):
     ''' Helper function to map a sequence of tokens into a sequence of
     AllenNLP Tokens, then wrap in a TextField with the given indexers '''
     return TextField(list(map(Token, sent)), token_indexers=indexers)
@@ -75,10 +73,10 @@ def process_single_pair_task_split(split, indexers, is_pair=True, classification
     '''
     def _make_instance(input1, input2, labels, idx):
         d = {}
-        d["input1"] = _sentence_to_text_field(input1, indexers)
+        d["input1"] = sentence_to_text_field(input1, indexers)
         d['sent1_str'] = MetadataField(" ".join(input1[1:-1]))
         if input2:
-            d["input2"] = _sentence_to_text_field(input2, indexers)
+            d["input2"] = sentence_to_text_field(input2, indexers)
             d['sent2_str'] = MetadataField(" ".join(input2[1:-1]))
         if classification:
             d["labels"] = LabelField(labels, label_namespace="labels",
@@ -245,334 +243,6 @@ class PairClassificationTask(ClassificationTask):
         return process_single_pair_task_split(split, indexers, is_pair=True)
 
 
-# SRL CoNLL 2005, formulated as an edge-labeling task.
-@register_task('edges-srl-conll2005', rel_path='edges/srl_conll2005',
-               label_file="labels.txt", files_by_split={
-                   'train': "train.edges.json",
-                   'val': "dev.edges.json",
-                   'test': "test.wsj.edges.json",
-               }, is_symmetric=False)
-# SRL CoNLL 2012 (OntoNotes), formulated as an edge-labeling task.
-@register_task('edges-srl-conll2012', rel_path='edges/srl_conll2012',
-               label_file="labels.txt", files_by_split={
-                   'train': "train.edges.json",
-                   'val': "dev.edges.json",
-                   'test': "test.edges.json",
-               }, is_symmetric=False)
-# SPR1, as an edge-labeling task (multilabel).
-@register_task('edges-spr1', rel_path='edges/spr1',
-               label_file="labels.txt", files_by_split={
-                   'train': "spr1.train.json",
-                   'val': "spr1.dev.json",
-                   'test': "spr1.test.json",
-               }, is_symmetric=False)
-# SPR2, as an edge-labeling task (multilabel).
-@register_task('edges-spr2', rel_path='edges/spr2',
-               label_file="labels.txt", files_by_split={
-                   'train': "train.edges.json",
-                   'val': "dev.edges.json",
-                   'test': "test.edges.json",
-               }, is_symmetric=False)
-# Definite pronoun resolution. Two labels.
-@register_task('edges-dpr', rel_path='edges/dpr',
-               label_file="labels.txt", files_by_split={
-                   'train': "train.edges.json",
-                   'val': "dev.edges.json",
-                   'test': "test.edges.json",
-               }, is_symmetric=False)
-# Coreference on OntoNotes corpus. Two labels.
-@register_task('edges-coref-ontonotes', rel_path='edges/ontonotes-coref',
-               label_file="labels.txt", files_by_split={
-                   'train': "train.edges.json",
-                   'val': "dev.edges.json",
-                   'test': "test.edges.json",
-               }, is_symmetric=False)
-# Re-processed version of the above, via AllenNLP data loaders.
-@register_task('edges-coref-ontonotes-conll',
-               rel_path='edges/ontonotes-coref-conll',
-               label_file="labels.txt", files_by_split={
-                   'train': "coref_conll_ontonotes_en_train.json",
-                   'val': "coref_conll_ontonotes_en_dev.json",
-                   'test': "coref_conll_ontonotes_en_test.json",
-               }, is_symmetric=False)
-# Entity type labeling on CoNLL 2003.
-@register_task('edges-ner-conll2003', rel_path='edges/ner_conll2003',
-               label_file="labels.txt", files_by_split={
-                   'train': "CoNLL-2003_train.json",
-                   'val': "CoNLL-2003_dev.json",
-                   'test': "CoNLL-2003_test.json",
-               }, single_sided=True)
-# Entity type labeling on OntoNotes.
-@register_task('edges-ner-ontonotes',
-               rel_path='edges/ontonotes-ner',
-               label_file="labels.txt", files_by_split={
-                   'train': "ner_ontonotes_en_train.json",
-                   'val': "ner_ontonotes_en_dev.json",
-                   'test': "ner_ontonotes_en_test.json",
-               }, single_sided=True)
-# Dependency edge labeling on UD treebank (GUM). Use 'ewt' version instead.
-@register_task('edges-dep-labeling', rel_path='edges/dep',
-               label_file="labels.txt", files_by_split={
-                   'train': "train.json",
-                   'val': "dev.json",
-                   'test': "test.json",
-               }, is_symmetric=False)
-# Dependency edge labeling on English Web Treebank (UD).
-@register_task('edges-dep-labeling-ewt', rel_path='edges/dep_ewt',
-               label_file="labels.txt", files_by_split={
-                   'train': "train.edges.json",
-                   'val': "dev.edges.json",
-                   'test': "test.edges.json",
-               }, is_symmetric=False)
-# PTB constituency membership / labeling.
-@register_task('edges-constituent-ptb', rel_path='edges/ptb-membership',
-               label_file="labels.txt", files_by_split={
-                   'train': "ptb_train.json",
-                   'val': "ptb_dev.json",
-                   'test': "ptb_test.json",
-               }, single_sided=True)
-# Constituency membership / labeling on OntoNotes.
-@register_task('edges-constituent-ontonotes',
-               rel_path='edges/ontonotes-constituents',
-               label_file="labels.txt", files_by_split={
-                   'train': "consts_ontonotes_en_train.json",
-                   'val': "consts_ontonotes_en_dev.json",
-                   'test': "consts_ontonotes_en_test.json",
-               }, single_sided=True)
-@register_task('edges-pos-ontonotes',
-               rel_path='edges/ontonotes-constituents',
-               label_file="labels.pos.txt", files_by_split={
-                   'train': "consts_ontonotes_en_train.pos.json",
-                   'val': "consts_ontonotes_en_dev.pos.json",
-                   'test': "consts_ontonotes_en_test.pos.json",
-               }, single_sided=True)
-@register_task('edges-nonterminal-ontonotes',
-               rel_path='edges/ontonotes-constituents',
-               label_file="labels.nonterminal.txt", files_by_split={
-                   'train': "consts_ontonotes_en_train.nonterminal.json",
-                   'val': "consts_ontonotes_en_dev.nonterminal.json",
-                   'test': "consts_ontonotes_en_test.nonterminal.json",
-               }, single_sided=True)
-# CCG tagging (tokens only).
-@register_task('edges-ccg-tag', rel_path='edges/ccg_tag',
-               label_file="labels.txt", files_by_split={
-                   'train': "ccg.tag.train.json",
-                   'val': "ccg.tag.dev.json",
-                   'test': "ccg.tag.test.json",
-               }, single_sided=True)
-# CCG parsing (constituent labeling).
-@register_task('edges-ccg-parse', rel_path='edges/ccg_parse',
-               label_file="labels.txt", files_by_split={
-                   'train': "ccg.parse.train.json",
-                   'val': "ccg.parse.dev.json",
-                   'test': "ccg.parse.test.json",
-               }, single_sided=True)
-class EdgeProbingTask(Task):
-    ''' Generic class for fine-grained edge probing.
-
-    Acts as a classifier, but with multiple targets for each input text.
-
-    Targets are of the form (span1, span2, label), where span1 and span2 are
-    half-open token intervals [i, j).
-
-    Subclass this for each dataset, or use register_task with appropriate kw
-    args.
-    '''
-    @property
-    def _tokenizer_suffix(self):
-        ''' Suffix to make sure we use the correct source files. '''
-        return ".retokenized." + self.tokenizer_name
-
-    def __init__(self, path: str, max_seq_len: int,
-                 name: str,
-                 label_file: str = None,
-                 files_by_split: Dict[str, str] = None,
-                 is_symmetric: bool = False,
-                 single_sided: bool = False):
-        """Construct an edge probing task.
-
-        path, max_seq_len, and name are passed by the code in preprocess.py;
-        remaining arguments should be provided by a subclass constructor or via
-        @register_task.
-
-        Args:
-            path: data directory
-            max_seq_len: maximum sequence length (currently ignored)
-            name: task name
-            label_file: relative path to labels file
-            files_by_split: split name ('train', 'val', 'test') mapped to
-                relative filenames (e.g. 'train': 'train.json')
-            is_symmetric: if true, span1 and span2 are assumed to be the same
-                type and share parameters. Otherwise, we learn a separate
-                projection layer and attention weight for each.
-            single_sided: if true, only use span1.
-        """
-        super().__init__(name)
-
-        assert label_file is not None
-        assert files_by_split is not None
-        self._files_by_split = {
-            split: os.path.join(path, fname) + self._tokenizer_suffix
-            for split, fname in files_by_split.items()
-        }
-        self._iters_by_split = self.load_data()
-        self.max_seq_len = max_seq_len
-        self.is_symmetric = is_symmetric
-        self.single_sided = single_sided
-
-        label_file = os.path.join(path, label_file)
-        self.all_labels = list(utils.load_lines(label_file))
-        self.n_classes = len(self.all_labels)
-        # see add_task_label_namespace in preprocess.py
-        self._label_namespace = self.name + "_labels"
-
-        # Scorers
-        #  self.acc_scorer = CategoricalAccuracy()  # multiclass accuracy
-        self.mcc_scorer = FastMatthews()
-        self.acc_scorer = BooleanAccuracy()  # binary accuracy
-        self.f1_scorer = F1Measure(positive_label=1)  # binary F1 overall
-        self.val_metric = "%s_f1" % self.name  # TODO: switch to MCC?
-        self.val_metric_decreases = False
-
-    def _stream_records(self, filename):
-        skip_ctr = 0
-        total_ctr = 0
-        for record in utils.load_json_data(filename):
-            total_ctr += 1
-            # Skip records with empty targets.
-            # TODO(ian): don't do this if generating negatives!
-            if not record.get('targets', None):
-                skip_ctr += 1
-                continue
-            yield record
-        log.info("Read=%d, Skip=%d, Total=%d from %s",
-                 total_ctr - skip_ctr, skip_ctr, total_ctr,
-                 filename)
-
-    @staticmethod
-    def merge_preds(record: Dict, preds: Dict) -> Dict:
-        """ Merge predictions into record, in-place.
-
-        List-valued predictions should align to targets,
-        and are attached to the corresponding target entry.
-
-        Non-list predictions are attached to the top-level record.
-        """
-        record['preds'] = {}
-        for target in record['targets']:
-            target['preds'] = {}
-        for key, val in preds.items():
-            if isinstance(val, list):
-                assert len(val) == len(record['targets'])
-                for i, target in enumerate(record['targets']):
-                    target['preds'][key] = val[i]
-            else:
-                # non-list predictions, attach to top-level preds
-                record['preds'][key] = val
-        return record
-
-    def load_data(self):
-        iters_by_split = collections.OrderedDict()
-        for split, filename in self._files_by_split.items():
-            #  # Lazy-load using RepeatableIterator.
-            #  loader = functools.partial(utils.load_json_data,
-            #                             filename=filename)
-            #  iter = serialize.RepeatableIterator(loader)
-            iter = list(self._stream_records(filename))
-            iters_by_split[split] = iter
-        return iters_by_split
-
-    def get_split_text(self, split: str):
-        ''' Get split text as iterable of records.
-
-        Split should be one of 'train', 'val', or 'test'.
-        '''
-        return self._iters_by_split[split]
-
-    def get_num_examples(self, split_text):
-        ''' Return number of examples in the result of get_split_text.
-
-        Subclass can override this if data is not stored in column format.
-        '''
-        return len(split_text)
-
-    def _make_span_field(self, s, text_field, offset=1):
-        return SpanField(s[0] + offset, s[1] - 1 + offset, text_field)
-
-    def make_instance(self, record, idx, indexers) -> Type[Instance]:
-        """Convert a single record to an AllenNLP Instance."""
-        tokens = record['text'].split()  # already space-tokenized by Moses
-        tokens = [utils.SOS_TOK] + tokens + [utils.EOS_TOK]
-        text_field = _sentence_to_text_field(tokens, indexers)
-
-        d = {}
-        d["idx"] = MetadataField(idx)
-
-        d['input1'] = text_field
-
-        d['span1s'] = ListField([self._make_span_field(t['span1'], text_field, 1)
-                                 for t in record['targets']])
-        if not self.single_sided:
-            d['span2s'] = ListField([self._make_span_field(t['span2'], text_field, 1)
-                                     for t in record['targets']])
-
-        # Always use multilabel targets, so be sure each label is a list.
-        labels = [utils.wrap_singleton_string(t['label'])
-                  for t in record['targets']]
-        d['labels'] = ListField([MultiLabelField(label_set,
-                                                 label_namespace=self._label_namespace,
-                                                 skip_indexing=False)
-                                 for label_set in labels])
-        return Instance(d)
-
-    def process_split(self, records, indexers) -> Iterable[Type[Instance]]:
-        ''' Process split text into a list of AllenNLP Instances. '''
-        def _map_fn(r, idx): return self.make_instance(r, idx, indexers)
-        return map(_map_fn, records, itertools.count())
-
-    def get_all_labels(self) -> List[str]:
-        return self.all_labels
-
-    def get_sentences(self) -> Iterable[Sequence[str]]:
-        ''' Yield sentences, used to compute vocabulary. '''
-        for split, iter in self._iters_by_split.items():
-            # Don't use test set for vocab building.
-            if split.startswith("test"):
-                continue
-            for record in iter:
-                yield record["text"].split()
-
-    def get_metrics(self, reset=False):
-        '''Get metrics specific to the task'''
-        metrics = {}
-        metrics['mcc'] = self.mcc_scorer.get_metric(reset)
-        metrics['acc'] = self.acc_scorer.get_metric(reset)
-        precision, recall, f1 = self.f1_scorer.get_metric(reset)
-        metrics['precision'] = precision
-        metrics['recall'] = recall
-        metrics['f1'] = f1
-        return metrics
-
-
-class OpenAIEdgeProbingTask(EdgeProbingTask):
-    """Version of EdgeProbingTask that loads BPE-tokenized data."""
-    @property
-    def tokenizer_name(self):
-        return "OpenAI.BPE"
-
-# We need '-openai' versions of all edge probing tasks, which load the correct
-# tokenization for that model. To avoid lots of boilerplate, add these to the
-# registry at import time using the loop below.
-for name in list(REGISTRY.keys()):
-    args = REGISTRY[name]
-    cls = args[0]
-    if (issubclass(cls, EdgeProbingTask)
-            and not issubclass(cls, OpenAIEdgeProbingTask)):
-        new_args = (OpenAIEdgeProbingTask, *args[1:])
-        new_name = name + "-openai"
-        REGISTRY[new_name] = new_args
-
-
 class PairRegressionTask(RegressionTask):
     ''' Generic sentence pair classification '''
 
@@ -716,9 +386,9 @@ class LanguageModelingTask(SequenceGenerationTask):
             to avoid issues with needing to strip extra tokens
             in the input for each direction '''
             d = {}
-            d["input"] = _sentence_to_text_field(sent, indexers)
-            d["targs"] = _sentence_to_text_field(sent[1:] + [sent[0]], self.target_indexer)
-            d["targs_b"] = _sentence_to_text_field([sent[-1]] + sent[:-1], self.target_indexer)
+            d["input"] = sentence_to_text_field(sent, indexers)
+            d["targs"] = sentence_to_text_field(sent[1:] + [sent[0]], self.target_indexer)
+            d["targs_b"] = sentence_to_text_field([sent[-1]] + sent[:-1], self.target_indexer)
             return Instance(d)
         for sent in split:
             yield _make_instance(sent)
@@ -877,9 +547,9 @@ class RedditTask(RankingTask):
         ''' Process split text into a list of AllenNLP Instances. '''
         def _make_instance(input1, input2, labels):
             d = {}
-            d["input1"] = _sentence_to_text_field(input1, indexers)
+            d["input1"] = sentence_to_text_field(input1, indexers)
             #d['sent1_str'] = MetadataField(" ".join(input1[1:-1]))
-            d["input2"] = _sentence_to_text_field(input2, indexers)
+            d["input2"] = sentence_to_text_field(input2, indexers)
             #d['sent2_str'] = MetadataField(" ".join(input2[1:-1]))
             d["labels"] = LabelField(labels, label_namespace="labels",
                                      skip_indexing=True)
@@ -951,9 +621,9 @@ class RedditPairClassificationTask(PairClassificationTask):
         ''' Process split text into a list of AllenNLP Instances. '''
         def _make_instance(input1, input2, labels):
             d = {}
-            d["input1"] = _sentence_to_text_field(input1, indexers)
+            d["input1"] = sentence_to_text_field(input1, indexers)
             #d['sent1_str'] = MetadataField(" ".join(input1[1:-1]))
-            d["input2"] = _sentence_to_text_field(input2, indexers)
+            d["input2"] = sentence_to_text_field(input2, indexers)
             #d['sent2_str'] = MetadataField(" ".join(input2[1:-1]))
             d["labels"] = LabelField(labels, label_namespace="labels",
                                      skip_indexing=True)
@@ -1486,8 +1156,8 @@ class MultiNLIDiagnosticTask(PairClassificationTask):
         def _make_instance(input1, input2, label, idx, lex_sem, pr_ar_str, logic, knowledge):
             ''' from multiple types in one column create multiple fields '''
             d = {}
-            d["input1"] = _sentence_to_text_field(input1, indexers)
-            d["input2"] = _sentence_to_text_field(input2, indexers)
+            d["input1"] = sentence_to_text_field(input1, indexers)
+            d["input2"] = sentence_to_text_field(input2, indexers)
             d["labels"] = LabelField(label, label_namespace="labels",
                                      skip_indexing=True)
             d["idx"] = LabelField(idx, label_namespace="idx",
@@ -1859,8 +1529,8 @@ class MTTask(SequenceGenerationTask):
         ''' Process split text into a list of AllenNLP Instances. '''
         def _make_instance(input, target):
             d = {}
-            d["inputs"] = _sentence_to_text_field(input, indexers)
-            d["targs"] = _sentence_to_text_field(target, self.target_indexer)  # this line changed
+            d["inputs"] = sentence_to_text_field(input, indexers)
+            d["targs"] = sentence_to_text_field(target, self.target_indexer)  # this line changed
             return Instance(d)
 
         for sent1, sent2 in split:
@@ -1992,8 +1662,8 @@ class Wiki103Classification(PairClassificationTask):
         ''' Process a language modeling split.  Split is a single list of sentences here.  '''
         def _make_instance(input1, input2, labels):
             d = {}
-            d["input1"] = _sentence_to_text_field(input1, indexers)
-            d["input2"] = _sentence_to_text_field(input2, indexers)
+            d["input1"] = sentence_to_text_field(input1, indexers)
+            d["input2"] = sentence_to_text_field(input2, indexers)
             d["labels"] = LabelField(labels, label_namespace="labels",
                                      skip_indexing=True)
             return Instance(d)
@@ -2061,8 +1731,8 @@ class Wiki103Seq2SeqTask(MTTask):
 
         def _make_instance(prev_sent, sent):
             d = {}
-            d["inputs"] = _sentence_to_text_field(prev_sent, indexers)
-            d["targs"] = _sentence_to_text_field(sent, target_indexer)
+            d["inputs"] = sentence_to_text_field(prev_sent, indexers)
+            d["targs"] = sentence_to_text_field(sent, target_indexer)
             return Instance(d)
 
         prev_sent = None
@@ -2131,8 +1801,8 @@ class DisSentTask(PairClassificationTask):
         ''' Process split text into a list of AllenNLP Instances. '''
         def _make_instance(input1, input2, labels):
             d = {}
-            d["input1"] = _sentence_to_text_field(input1, indexers)
-            d["input2"] = _sentence_to_text_field(input2, indexers)
+            d["input1"] = sentence_to_text_field(input1, indexers)
+            d["input2"] = sentence_to_text_field(input2, indexers)
             d["labels"] = LabelField(labels, label_namespace="labels",
                                      skip_indexing=True)
             return Instance(d)
@@ -2238,7 +1908,7 @@ class GroundedTask(Task):
         Returns:
         '''
         def _make_instance(sent, label, ids):
-            input1 = _sentence_to_text_field(sent, indexers)
+            input1 = sentence_to_text_field(sent, indexers)
             label = NumericField(label)
             ids = NumericField(ids)
             return Instance({"input1": input1, "labels": label, "ids": ids})
@@ -2348,7 +2018,7 @@ class GroundedSWTask(Task):
         Returns:
         '''
         def _make_instance(sent, label, ids):
-            input1 = _sentence_to_text_field(sent, indexers)
+            input1 = sentence_to_text_field(sent, indexers)
             label = NumericField(label)
             ids = NumericField(ids)
             return Instance({"input1": input1, "labels": label, "ids": ids})
